fix telegram_id to bigInteger, disable timestamp for User model, docker fix, setting up handle action of TelegramController

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('telegram_id')->unique();
+            $table->bigInteger('telegram_id')->unique();
             $table->boolean('subscribed');
         });
